// app/src/ui/components/atomicDesign/organisms/forecasts/Warning/__generated__/WarningFragment.graphql.ts
export type WarningFragment$data = {
  readonly query: {
    readonly Warning_Claims: {
      readonly edges: ReadonlyArray<{
        readonly node: {
          readonly Acc_CostCategory__c: {
            readonly value: string | null;
          } | null;
          readonly Acc_PeriodCostCategoryTotal__c: {
            readonly value: number | null;
          } | null;
          readonly Acc_ProjectPeriodNumber__c: {
            readonly value: number | null;
          } | null;
          readonly Id: string;
          readonly RecordType: {
            readonly Name: {
              readonly value: string | null;
            } | null;
          } | null;
        } | null;
      } | null> | null;
    } | null;
    readonly Warning_CostCategory: {
      readonly edges: ReadonlyArray<{
        readonly node: {
          readonly Acc_CostCategoryName__c: {
            readonly value: string | null;
          } | null;
          readonly Acc_DisplayOrder__c: {
            readonly value: number | null;
          } | null;
          readonly Id: string;
        } | null;
      } | null> | null;
    } | null;
    readonly Warning_Profile: {
      readonly edges: ReadonlyArray<{
        readonly node: {
          readonly Acc_CostCategoryGOLCost__c: {
            readonly value: number | null;
          } | null;
          readonly Acc_CostCategory__c: {
            readonly value: string | null;
          } | null;
          readonly Acc_LatestForecastCost__c: {
            readonly value: number | null;
          } | null;
          readonly Acc_ProjectPeriodNumber__c: {
            readonly value: number | null;
          } | null;
          readonly Id: string;
          readonly RecordType: {
            readonly Name: {
              readonly value: string | null;
            } | null;
          } | null;
        } | null;
      } | null> | null;
    } | null;
    readonly Warning_Project: {
      readonly edges: ReadonlyArray<{
        readonly node: {
          readonly Acc_ProjectParticipantsProject__r: {
            readonly edges: ReadonlyArray<{
              readonly node: {
                readonly Acc_AccountId__c: {
                  readonly value: string | null;
                } | null;
                readonly Id: string;
              } | null;
            } | null> | null;
          } | null;
          readonly Id: string;
          readonly roles: {
            readonly isFc: boolean;
            readonly isMo: boolean;
            readonly isPm: boolean;
            readonly partnerRoles: ReadonlyArray<{
              readonly isFc: boolean;
              readonly isMo: boolean;
              readonly isPm: boolean;
              readonly partnerId: string;
            }>;
          };
        } | null;
      } | null> | null;
    } | null;
  };
  readonly " $fragmentType": "WarningFragment";
};
export type WarningFragment$key = {
  readonly " $data"?: WarningFragment$data;
  readonly " $fragmentSpreads": FragmentRefs<"WarningFragment">;
};